Refs #24. Pull out test_util tests into a test test_util
